# Conflicts:
#	app/src/main/res/values/dimens.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/bink/wallet/model/response/membership_card/CardBalance.kt
@@ -11,5 +11,5 @@ data class CardBalance(
     var curency: String?,
     var prefix: String?,
     var suffix: String?,
-    var updated_at: Int?
+    var updated_at: Long?
 ) : Parcelable
